handle dataset-specific overrides to useCasesExcluded
