+++ server/routes/termdb.config.ts
@@ -102,6 +102,7 @@ function make(q, res, ds: Mds3WithCohort, genome) {
 	if (tdb.displaySampleIds) c.displaySampleIds = tdb.displaySampleIds
 	if (tdb.hasAncestry) c.hasAncestry = tdb.hasAncestry
 	if (tdb.logscaleBase2) c.logscaleBase2 = tdb.logscaleBase2
+	if (tdb.useCasesExcluded) c.useCasesExcluded = tdb.useCasesExcluded
 
 	if (ds.assayAvailability) c.assayAvailability = ds.assayAvailability
 	if (ds.customTwQByType) c.customTwQByType = ds.customTwQByType
